ENH: huge refactor for new error handling
New error handling for samplers, checkers and generators. Now error
messages are more informative and actually say what's broken and where
(e.g. RE in program, error in generator, ...)
The whole error handling has been rewritten.

One breaking change is that for now if custom checker wants to display a
message on failure, it has to be written to stderr.